--- src/renderer/features/proxy-remove-pure/index.ts
@@ -1,7 +1,0 @@
-import { removePureProxyModel } from './model/remove-pure-proxy-model';
-import { RemovePureProxy } from './ui/RemovePureProxy';
-
-export const proxyRemovePureFeature = {
-  views: { RemovePureProxy },
-  models: { removePureProxy: removePureProxyModel },
-};
