+++ src/renderer/features/proxy-remove/index.ts
@@ -1,11 +1,7 @@
 import { removeProxyModel } from './model/remove-proxy-model';
-import { RemoveProxy } from './ui/RemoveProxy';
+import { RemovePureProxy } from './ui/RemoveProxy';
 
-export const proxyRemoveFeature = {
-  views: {
-    RemoveProxy,
-  },
-  models: {
-    removeProxy: removeProxyModel,
-  },
+export const proxyRemovePureFeature = {
+  views: { RemovePureProxy },
+  models: { removePureProxy: removeProxyModel },
 };
